Create 0x02-python-import_modules
Write a program that imports the function def add(a, b): from the file add_0.py and prints the result of the addition 1 + 2 = 3

You have to use print function with string format to display integers You have to assign: the value 1 to a variable called a the value 2 to a variable called b and use those two variables as arguments when calling the functions add and print a and b must be defined in 2 different lines: a = 1 and another b = 2 Your program should print: + = <add(a, b) value> followed with a new line You can only use the word add_0 once in your code You are not allowed to use * for importing or import Your code should not be executed when imported - by using import, like the example below Write a program that imports functions from the file calculator_1.py, does some Maths, and prints the result. Write a program that prints the number of and the list of its arguments. Write a program that prints the result of the addition of all arguments

The output should be the result of the addition of all arguments, followed by a new line You can cast arguments into integers by using int() (you can assume that all arguments can be casted into integers) Your code should not be executed when imported Write a program that prints all the names defined by the compiled module hidden_4.pyc (please download it locally). Write a program that imports the variable a from the file variable_load_5.py and prints its value.

You are not allowed to use * for importing or import Your code should not be executed when imported Write a program that imports all functions from the file calculator_1.py and handles basic operations. Write a program that prints #pythoniscool, followed by a new line, in the standard output Write the Python function def magic_calculation(a, b): that does exactly the same as the following Python bytecode: Write a program that prints the alphabet in uppercase, followed by a new line.